+++ GU2_Allswellhospital/Controllers/WardManagementController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Data;
 using System.Data.Entity;
+using System.Data.Entity.Migrations;
 using System.Linq;
 using System.Net;
 using System.Web;
@@ -23,7 +24,27 @@ namespace GU2_Allswellhospital.Controllers
         // GET: WardManagement
         public ActionResult Index()
         {
-            return View(db.Wards.ToList());
+            //initalises all wards with spaces taken, if a patient is admitted to that ward
+            List<Patient> patients = db.Patients.Include(w => w.Ward).ToList();
+            List<Ward> wards = db.Wards.ToList();
+
+            //scans through each ward using a nested for each patient to find patient who are indeed admitted to that ward
+            foreach (Ward w in wards)
+            {
+
+                foreach(Patient p in patients)
+                {
+
+                    if(p.WardNo == w.WardNo)
+                    {
+                        w.WardSpacesTaken = w.WardSpacesTaken + 1;
+                    }
+
+                }
+
+            }
+
+            return View(wards);
         }
 
         // GET: WardManagement/Details/5
@@ -54,13 +75,36 @@ namespace GU2_Allswellhospital.Controllers
         [ValidateAntiForgeryToken]
         public ActionResult Create([Bind(Include = "WardNo,WardName,WardCapacity")] Ward ward)
         {
+            //checks if model submitted is valid
             if (ModelState.IsValid)
             {
+                if (!(ward.WardCapacity > 0))
+                {
+                    ViewBag.ErrorMessage = "Ward must have at least one space";
+                    return View(ward);
+                }
+
+                //temp listing of wards
+                List<Ward> wards = db.Wards.ToList();
+
+                //checks if any occurance of this was already exists
+                foreach (Ward w in wards)
+                {
+                    if (w.WardName == ward.WardName)
+                    {
+                        ViewBag.ErrorMessage = "Ward Already exsists, please choose different name";
+                        return View(ward);
+                    }
+                }
+
+                //if none already exists then added
                 db.Wards.Add(ward);
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
 
+            //invalid submittion message
+            ViewBag.ErrorMessage = "Submit is invalid, please fill all fields";
             return View(ward);
         }
 
@@ -88,10 +132,36 @@ namespace GU2_Allswellhospital.Controllers
         {
             if (ModelState.IsValid)
             {
-                db.Entry(ward).State = EntityState.Modified;
+
+                if (!(ward.WardCapacity > 0))
+                {
+                    ViewBag.ErrorMessage = "Ward must have at least one space";
+                    return View(ward);
+                }
+
+                //temp listing of wards
+                List<Ward> wards = db.Wards.ToList();
+
+                //checks if any occurance of this was already exists
+                foreach (Ward w in wards)
+                {
+                    if(w.WardNo == ward.WardNo)
+                    {
+                        continue;
+                    }
+
+                    if (w.WardName == ward.WardName)
+                    {
+                        ViewBag.ErrorMessage = "Ward Already exsists, please choose different name";
+                        return View(ward);
+                    }
+                }
+
+                db.Set<Ward>().AddOrUpdate(ward);
                 db.SaveChanges();
                 return RedirectToAction("Index");
             }
+            ViewBag.ErrorMessage = "Submittion failed, model invalid";
             return View(ward);
         }
 
@@ -116,6 +186,16 @@ namespace GU2_Allswellhospital.Controllers
         public ActionResult DeleteConfirmed(string id)
         {
             Ward ward = db.Wards.Find(id);
+
+            List<Admission> admissions = db.Admissions.Include(a => a.Ward).Where(a => a.isAdmitted == true).ToList();
+
+            //checks is any admission still exists to this ward
+            if(admissions.Count > 0)
+            {
+                ViewBag.ErrorMessage = "Patients are still assigned to this ward";
+                return View(ward);
+            }
+
             db.Wards.Remove(ward);
             db.SaveChanges();
             return RedirectToAction("Index");
